test/make-binary-wrapper: sync with nixpkgs
