+++ test/make-binary-wrapper/add-flags/add-flags.env
@@ -4,7 +4,11 @@ SUBST_ARGV0
 -y
 -z
 -abc
+test var here
 -g
 *.txt
+-a
+*
 -foo
 -bar
+test var 2 here
