+++ test/make-binary-wrapper/argv0/argv0.env
