+++ test/make-binary-wrapper/chdir/chdir.cmdline
