+++ test/make-binary-wrapper/chdir/chdir.env
