+++ test/make-binary-wrapper/combination/combination.cmdline
