+++ test/make-binary-wrapper/inherit-argv0/inherit-argv0.cmdline
