+++ test/make-binary-wrapper/invalid-env/invalid-env.c
