+++ test/make-binary-wrapper/overlength-strings/overlength-strings.cmdline
