+++ test/make-binary-wrapper/prefix/prefix.cmdline
