+++ test/make-binary-wrapper/suffix/suffix.cmdline
